# backend/openapi.yaml
openapi: 3.0.3
info:
  title: DotBot Backend API
  description: |
    DotBot Backend API - AI-powered Polkadot blockchain operations through natural language.
    
    This API provides endpoints for:
    - Chat interactions with AI (via `/api/chat`)
    - Full DotBot functionality with blockchain operations (via `/api/dotbot`)
    - Session management for DotBot instances
    
    All AI communication happens on the backend where API keys are securely stored.
  version: 0.1.0
  contact:
    name: DotBot Team
  license:
    name: GNU General Public License v3.0
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.dotbot.example.com
    description: Production server (example)
...
paths:
...
  /api/chat:
    post:
      tags:
        - Chat
      summary: Send a chat message
      description: Simple AI chat endpoint that processes messages using the configured AI provider
      operationId: sendChatMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/dotbot/chat:
    post:
      tags:
        - DotBot
      summary: Send a DotBot chat message
      description: |
        Full DotBot chat endpoint that handles AI communication and blockchain operations.
        This endpoint processes natural language requests and can execute blockchain transactions.
      operationId: sendDotBotMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DotBotChatRequest'
      responses:
        '200':
          description: Chat result with execution information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DotBotChatResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/dotbot/session/{sessionId}/chats:
    get:
      tags:
        - Chat Instances
      summary: List all chat instances
      description: Retrieves all chat instances for a specific session
      operationId: listChatInstances
      parameters:
        - name: sessionId
          in: path
          required: true
          description: Session identifier
          schema:
            type: string
            example: wallet:5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY:mainnet
      responses:
        '200':
          description: List of chat instances
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  chats:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatInstanceData'
                  timestamp:
                    type: string
                    format: date-time
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/dotbot/session/{sessionId}/chats/{chatId}:
    get:
      tags:
        - Chat Instances
      summary: Get chat instance
      description: Retrieves a specific chat instance
      operationId: getChatInstance
      parameters:
        - name: sessionId
          in: path
          required: true
          description: Session identifier
          schema:
            type: string
        - name: chatId
          in: path
          required: true
          description: Chat instance identifier
          schema:
            type: string
      responses:
        '200':
          description: Chat instance data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  chat:
                    $ref: '#/components/schemas/ChatInstanceData'
                  timestamp:
                    type: string
                    format: date-time
        '404':
          description: Chat instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Chat Instances
      summary: Delete chat instance
      description: Deletes a specific chat instance
      operationId: deleteChatInstance
      parameters:
        - name: sessionId
          in: path
          required: true
          description: Session identifier
          schema:
            type: string
        - name: chatId
          in: path
          required: true
          description: Chat instance identifier
          schema:
            type: string
      responses:
        '200':
          description: Deletion result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Chat instance deleted
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ChatRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: The chat message to send
          example: What is Polkadot?
        context:
          type: object
          description: Additional context for the AI
          additionalProperties: true
        provider:
          type: string
          enum: [asi-one, claude, openai]
          description: AI provider to use (optional, uses default if not specified)
          example: asi-one
    ChatResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        response:
          type: string
          description: AI response text
          example: Polkadot is a blockchain platform...
        provider:
          type: string
          example: asi-one
        timestamp:
          type: string
          format: date-time
    WalletAccount:
      type: object
      required:
        - address
        - source
      properties:
        address:
          type: string
          description: Polkadot wallet address (SS58 format)
          example: 5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY
        name:
          type: string
          description: Wallet account name (optional)
          example: Alice
        source:
          type: string
          description: Wallet source/provider
          example: polkadot-js
    ConversationMessage:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum: [user, assistant, system]
          description: Message role
        content:
          type: string
          description: Message content
        timestamp:
          type: number
          description: Unix timestamp in milliseconds
          example: 1705168800000
    DotBotChatRequest:
      type: object
      required:
        - message
        - wallet
      properties:
        message:
          type: string
          description: Natural language message for DotBot
          example: Transfer 1 DOT to Alice
        sessionId:
          type: string
          description: Optional session identifier. If not provided, will be generated from wallet address and environment.
          example: wallet:5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY:mainnet
        wallet:
          $ref: '#/components/schemas/WalletAccount'
        environment:
          type: string
          enum: [mainnet, testnet]
          default: mainnet
          description: Blockchain environment
          example: mainnet
        network:
          type: string
          enum: [polkadot, kusama, westend]
          description: |
            Blockchain network. If provided, environment will be derived from it.
            - polkadot, kusama → mainnet
            - westend → testnet
          example: polkadot
        options:
          type: object
          properties:
            systemPrompt:
              type: string
              description: Custom system prompt override
            conversationHistory:
              type: array
              items:
                $ref: '#/components/schemas/ConversationMessage'
              description: Previous conversation messages for context
            executionOptions:
              type: object
              description: Execution options (continueOnError, etc.)
              additionalProperties: true
        provider:
          type: string
          enum: [asi-one, claude, openai]
          description: AI provider to use
          example: asi-one
    ExecutionPlan:
      type: object
      properties:
        id:
          type: string
          description: Plan identifier
        steps:
          type: array
          items:
            type: object
            properties:
              agentClassName:
                type: string
                example: AssetTransferAgent
              functionName:
                type: string
                example: transfer
              parameters:
                type: object
                additionalProperties: true
              description:
                type: string
              executionType:
                type: string
        requiresApproval:
          type: boolean
    ChatResult:
      type: object
      properties:
        response:
          type: string
          description: LLM response text
          example: I'll transfer 1 DOT to Alice for you.
        plan:
          $ref: '#/components/schemas/ExecutionPlan'
        executed:
          type: boolean
          description: Whether operations were executed
          example: true
        success:
          type: boolean
          description: Execution success status
          example: true
        completed:
          type: number
          description: Number of operations completed
          example: 1
        failed:
          type: number
          description: Number of operations failed
          example: 0
    DotBotChatResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        result:
          $ref: '#/components/schemas/ChatResult'
        sessionId:
          type: string
          description: Session identifier used for this request
          example: wallet:5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY:mainnet
        chatId:
          type: string
          nullable: true
          description: Current chat instance ID (if any)
          example: chat_1705168800000_abc123
        timestamp:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Internal server error
        message:
          type: string
          description: Detailed error message
          example: Failed to process chat request
        timestamp:
          type: string
          format: date-time
          example: 2026-01-13T18:00:00.000Z
    ChatInstanceData:
      type: object
      properties:
        id:
          type: string
          description: Chat instance identifier
          example: chat_1705168800000_abc123
        name:
          type: string
          description: Display name for the chat
          example: Transfer DOT to Alice
        environment:
          type: string
          enum: [mainnet, testnet]
          description: Blockchain environment
          example: mainnet
        network:
          type: string
          enum: [polkadot, kusama, westend]
          description: Blockchain network
          example: polkadot
        walletAddress:
          type: string
          description: Associated wallet address
          example: 5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY
        createdAt:
          type: number
          description: Creation timestamp (Unix ms)
          example: 1705168800000
        lastModified:
          type: number
          description: Last modification timestamp (Unix ms)
          example: 1705169400000
        messageCount:
          type: integer
          description: Number of messages in the chat
          example: 5
        isEmpty:
          type: boolean
          description: Whether the chat is empty
          example: false
...